Improve remote shell command
